requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django>=5.0.1,<5.1.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.3.1,<4.4.0
psycopg2-binary>=2.9.9,<2.10.0
python-dotenv>=1.0.0,<1.1.0
django-filter>=23.5,<23.6
djangorestframework-simplejwt>=5.3.1,<5.4.0
flake8==7.2.0
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
sqlparse==0.5.3
drf-spectacular>=0.27.1,<0.28.0
